ENG-119 Move addArticleLinks to SearchController
