+++ src/test/java/org/ambraproject/wombat/controller/SearchControllerTest.java
@@ -25,19 +25,23 @@ package org.ambraproject.wombat.controller;
 import org.ambraproject.wombat.config.TestSpringConfiguration;
 import org.ambraproject.wombat.config.site.Site;
 import org.ambraproject.wombat.config.site.SiteSet;
+import org.ambraproject.wombat.service.remote.SolrSearchApi;
 import org.ambraproject.wombat.service.remote.SolrSearchApiImpl;
 import org.ambraproject.wombat.util.MockSiteUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.AbstractJUnit4SpringContextTests;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -66,5 +70,42 @@ public class SearchControllerTest extends AbstractJUnit4SpringContextTests {
     assertEquals("subject1", commonParams.subjectList.get(0));
     assertEquals("subject2", commonParams.subjectList.get(1));
     assertTrue(commonParams.isFiltered);
+  }
+
+  private class MySearchController extends SearchController {
+
+    protected void initializeEIssnToJournalKeyMap(SiteSet siteSet, Site currentSite)
+        throws IOException {
+      ImmutableMap.Builder<String, String> builder = new ImmutableMap.Builder<>();
+      builder.put("123", "journal1Key").put("456", "journal2Key").put("789",
+          "collectionJournalKey");
+      this.eIssnToJournalKey = builder.build();
+    }
+  }    
+  @Test
+  public void testAddArticleLinks() throws IOException {
+    SearchController searchController = new MySearchController();
+    List<Map<String,Object>> docs = new ArrayList<>(1);
+    Map<String,Object> doc = new HashMap<>();
+    List<String> crossPubbedJournals = new ArrayList<>(1);
+    crossPubbedJournals.add("journal1Key");
+    doc.put("id", "12345");
+    doc.put("eissn", "123");
+    docs.add(doc);
+    SolrSearchApi.Result searchResults =
+      SolrSearchApi.Result.builder()
+      .setDocs(docs)
+      .setNumFound(1)
+      .setStart(0)
+      .build();
+    MockHttpServletRequest request = new MockHttpServletRequest();
+    request.setContextPath("someContextPath");
+    Site site = MockSiteUtil.getByUniqueJournalKey(siteSet, "journal2Key");
+
+    List<Map<String, Object>> actualDocs = searchController.addArticleLinks(searchResults, request, site, siteSet).getDocs();
+    assertEquals(1, actualDocs.size());
+    Map<String, Object> actualDoc = actualDocs.get(0);
+    assertEquals("12345", actualDoc.get("id"));
+    assertTrue(actualDoc.get("link").toString().endsWith("someContextPath/site1/article?id=12345"));
   }
 }
