+++ src/test/java/org/ambraproject/wombat/service/remote/SolrSearchApiTest.java
@@ -233,45 +233,6 @@ public class SolrSearchApiTest extends AbstractJUnit4SpringContextTests {
     assertSubjects(actualMap.get("fq"), "\"Skull\"", "\"Head\"", "\"Teeth\"");
   }
 
-  private static class SearchApiForAddArticleLinksTest extends SolrSearchApiImpl {
-
-    @Override
-    protected void initializeEIssnToJournalKeyMap(SiteSet siteSet, Site currentSite) throws IOException {
-      ImmutableMap.Builder<String, String> builder = new ImmutableMap.Builder<>();
-      builder.put("123", "journal1Key")
-          .put("456", "journal2Key")
-          .put("789", "collectionJournalKey");
-      eIssnToJournalKey = builder.build();
-    }
-  }
-
-  @Test
-  public void testAddArticleLinks() throws IOException {
-    SolrSearchApi solrSearchApiForTest = new SearchApiForAddArticleLinksTest();
-    List<Map<String,Object>> docs = new ArrayList<>(1);
-    Map<String,Object> doc = new HashMap<>();
-    List<String> crossPubbedJournals = new ArrayList<>(1);
-    crossPubbedJournals.add("journal1Key");
-    doc.put("id", "12345");
-    doc.put("eissn", "123");
-    docs.add(doc);
-    SolrSearchApi.Result searchResults =
-      SolrSearchApi.Result.builder()
-      .setDocs(docs)
-      .setNumFound(1)
-      .setStart(0)
-      .build();
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    request.setContextPath("someContextPath");
-    Site site = MockSiteUtil.getByUniqueJournalKey(siteSet, "journal2Key");
-
-    List<Map<String, Object>> actualDocs = solrSearchApiForTest.addArticleLinks(searchResults, request, site, siteSet).getDocs();
-    assertEquals(1, actualDocs.size());
-    Map<String, Object> actualDoc = actualDocs.get(0);
-    assertEquals("12345", actualDoc.get("id"));
-    assertTrue(actualDoc.get("link").toString().endsWith("someContextPath/site1/article?id=12345"));
-  }
-
   @Test
   public void testBuildSearchClause() {
     assertEquals("subject:\"foo\"", SolrQueryBuilder.buildAndSearchClause("subject", Arrays.asList("foo")));
